+++ R/popmaps2-package.R
@@ -1,7 +1,6 @@
 #' popmaps2: Estimate and modernize ancestry probability surfaces
 #'
 #' @keywords internal
-#' @importFrom foreach %:% %dopar%
 #' @importFrom graphics boxplot mtext par plot points
 #' @importFrom grDevices col2rgb grey.colors rgb
 #' @importFrom methods as